fix airplane mode that did not work when there were no -remove options in dist.ini
This never showed up as broken tests because all unit tests use the
-remove option to strip out network- and git-using plugins. We push
those plugins into the '-remove' payload during BUILD, but PluginRemover
short-circuits if there was no '-remove' field at all in the payload
before the bundle is instantiated. (If it existed, the full element list
is still used, because the array reference doesn't change -- undef vs.
defined is the only important distinction.)

Consequently, we need to take care of those plugins ourselves -- which
we used to do until the changes in version 0.127 (commit 6b4fd05252).